pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector-sentiment-search"
version = "0.1.0"
description = "Advanced Qdrant vector database implementation for sentiment analysis"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11.8"
license = {text = "MIT"}
keywords = ["qdrant", "vector-database", "embeddings", "sentiment-analysis", "nlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.2",
    "fastembed>=0.2.0",  # Sparse vector embeddings (SPLADE)
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",  # K-means clustering & metrics
    "pyyaml>=6.0",  # YAML config file parsing
    "faker>=20.0.0",  # Synthetic data generation
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "types-requests>=2.31.0",
    "pandas-stubs>=2.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
vector_sentiment = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "PTH",    # flake8-use-pathlib
]
ignore = [
    "ANN101",  # Missing type annotation for self
    "ANN102",  # Missing type annotation for cls
    "S101",    # Use of assert
    "S311",    # Standard pseudo-random generators
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py
"tests/*" = ["S101", "ANN"]  # Allow assert in tests, skip annotations

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Changed from true - too strict for this project
check_untyped_defs = true
disallow_any_unimported = false  # Changed from true - third-party libs
disallow_any_expr = false  # Changed from true - too restrictive
disallow_any_explicit = false  # Changed from true - allows Any when needed
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_reexport = false  # Changed from true - allows easier imports
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "sentence_transformers.*",
    "pyarrow.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/vector_sentiment --cov-report=term-missing"
